fix: export ambiguity
